refactor: reorder page flow (#35)
* refactor: reorder page flow

* fix: give Link component just one child

* feat: add Learning Journey page
