+++ client/pages/learning-journey.tsx
@@ -1,0 +1,3 @@
+const LearningJourneyPage = () => {};
+
+export default LearningJourneyPage;
