+++ client/styles/Home.module.css
@@ -83,6 +83,7 @@
     border-radius: 10px;
     transition: color 0.15s ease, border-color 0.15s ease;
     max-width: 300px;
+    cursor: pointer;
 }
 
 .card:hover,
